+++ packages/fabric/editor-core/examples/3-collab.tsx
@@ -8,9 +8,13 @@ import { akColorN80 } from '@atlaskit/util-shared-styles';
 import Editor from './../src/editor';
 import EditorContext from './../src/editor/ui/EditorContext';
 import WithEditorActions from './../src/editor/ui/WithEditorActions';
-import { storyMediaProviderFactory } from '@atlaskit/editor-test-helpers';
+import {
+  storyMediaProviderFactory,
+  storyContextIdentifierProviderFactory,
+} from '@atlaskit/editor-test-helpers';
 import { storyData as mentionStoryData } from '@atlaskit/mention/dist/es5/support';
 import { storyData as emojiStoryData } from '@atlaskit/emoji/dist/es5/support';
+import { storyData as taskDecisionStoryData } from '@atlaskit/task-decision/dist/es5/support';
 
 import {
   akEditorCodeBackground,
@@ -120,6 +124,10 @@ export default function Example() {
               mentionProvider={Promise.resolve(
                 mentionStoryData.resourceProvider,
               )}
+              taskDecisionProvider={Promise.resolve(
+                taskDecisionStoryData.getMockTaskDecisionResource(),
+              )}
+              contextIdentifierProvider={storyContextIdentifierProviderFactory()}
               collabEditProvider={collabEditProvider('rick')}
               placeholder="Write something..."
               shouldFocus={false}